gui and logic separated
