+++ Applications/src/Calculator.java
@@ -5,8 +5,6 @@
  */
 
 import java.awt.*;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.StringSelection;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
@@ -19,18 +17,14 @@ import javax.swing.border.EtchedBorder;
 public class Calculator extends JFrame implements ActionListener, KeyListener{
     
     private JLabel resultField;
-    private Font font = new Font("Arial", Font.BOLD, 20);
 
     private HashMap<JButton, String> numberButtons;
     private HashMap<JButton, String> operatorButtons;
+    private JButton clear;
+    private Calculations calculations;
 
-    private String buffer1 = "0";
-    private String buffer2 = "";
-    private String operator = "";
-    private boolean operatorSet = false;
-    private boolean dotSet1 = false;
-    private boolean dotSet2 = false;;
-    private String curResult = "";
+    private Font font = new Font("Arial", Font.BOLD, 20);
+
 
     public Calculator(){
 
@@ -51,6 +45,8 @@ public class Calculator extends JFrame implements ActionListener, KeyListener{
 
         numberButtons = new HashMap<>();
         operatorButtons = new HashMap<>();
+
+        calculations = new Calculations(this);
 
         // Buttons of calculator initialization
         {
@@ -102,7 +98,7 @@ public class Calculator extends JFrame implements ActionListener, KeyListener{
             operatorButtons.put(multiply, "*");
             addComponentToContainer(multiply, container, 3, 3, 1, 1);
 
-            JButton clear = buttonInit("CE");
+            clear = buttonInit("CE");
             clear.setMargin(new Insets(0, 0, 0, 0));
             addComponentToContainer(clear, container, 0, 4, 1, 1);
 
@@ -152,203 +148,25 @@ public class Calculator extends JFrame implements ActionListener, KeyListener{
 
         JButton button=(JButton) event.getSource();
         if (numberButtons.containsKey(button)){
-            actionsByButtons(numberButtons.get(button));
+            calculations.actionsByButtons(numberButtons.get(button));
         } else if (operatorButtons.containsKey(button)){
-            actionsByButtons(operatorButtons.get(button));
-        } else if (button.getText().equals("CE")){
-            actionsByButtons("CE");
+            calculations.actionsByButtons(operatorButtons.get(button));
+        } else if (button.equals(clear)){
+            calculations.actionsByButtons("CE");
         }
     }
 
-    private void actionsByButtons(String buttonValue){
-        if (buttonValue.equals("CE")){
-            clearData();
-            resultField.setText(buffer1);
-        }
 
-        if (buffer2.equals("")) {
-            if (!operatorSet && numberButtons.containsValue(buttonValue) && curResult.equals("")) {
-                buffer1 = addNumbersBuffer1(buttonValue);
-                resultField.setText(buffer1);
-            } else if (!operatorSet && numberButtons.containsValue(buttonValue) && !curResult.equals("")){
-                clearData();
-                buffer1 = addNumbersBuffer1(buttonValue);
-                resultField.setText(buffer1);
-            }
-            if (!operatorSet && operatorButtons.containsValue(buttonValue)) {
-                if (buttonValue.equals("=")) {
-                    curResult = calculate(buffer1, "", "");
-                    copyCurResultToClipboard(curResult);
-                    resultField.setText(curResult);
-                    buffer1 = curResult;
-                    operatorSet = true;
-                    operator = "=";
-                } else {
-                    operator = buttonValue;
-                    operatorSet = true;
-                }
-            } else if (operatorSet && operatorButtons.containsValue(buttonValue)) {
-                if (buttonValue.equals("=")) {
-                    curResult=calculate(buffer1, "", "");
-                    copyCurResultToClipboard(curResult);
-                    resultField.setText(curResult);
-                    buffer1=curResult;
-                    operator = "=";
-                    operatorSet = true;
-                } else {
-                    operator = buttonValue;
-                }
-            } else if(operatorSet && numberButtons.containsValue(buttonValue) && !curResult.equals("")) {
-                if (operator.equals("=")) {
-                    clearData();
-                    buffer1=addNumbersBuffer1(buttonValue);
-                    resultField.setText(buffer1);
-                } else {
-                    buffer2 = addNumbersBuffer2(buttonValue);
-                    resultField.setText(buffer2);
-                }
-            } else if (operatorSet && numberButtons.containsValue(buttonValue)){
-                buffer2 = addNumbersBuffer2(buttonValue);
-                resultField.setText(buffer2);
-            }
-        } else if (numberButtons.containsValue(buttonValue)){
-            buffer2 = addNumbersBuffer2(buttonValue);
-            resultField.setText(buffer2);
-        } else if (operatorButtons.containsValue(buttonValue)){
-            if (!buttonValue.equals("=")){
-                curResult = calculate(buffer1 , operator, buffer2);
-                copyCurResultToClipboard(curResult);
-                resultField.setText(curResult);
-                buffer1 = curResult;
-                operator = buttonValue;
-                operatorSet = true;
-            } else {
-                curResult = calculate(buffer1 , operator, buffer2);
-                copyCurResultToClipboard(curResult);
-                resultField.setText(curResult);
-                buffer1 = curResult;
-                if (buffer1.equals("error")) {
-                    clearData();
-                }
-            }
-        }
-    }
-
-    private String addNumbersBuffer2(String curString){
-        if (buffer2.equals("") && (!dotSet2)) {
-            if (curString.equals(".")) {
-                buffer2 = "0" + curString;
-                dotSet2 = true;
-            } else {
-                buffer2 = curString;
-            }
-        } else if (!buffer2.equals("") && (!dotSet2)) {
-            if (curString.equals("0") && buffer2.equals("0")){
-                buffer2 = curString;
-            } else if (curString.equals(".")) {
-                buffer2 += curString;
-                dotSet2 = true;
-            } else {
-                buffer2 += curString;
-            }
-        } else if (!buffer2.equals("") && dotSet2) {
-            if (!curString.equals(".")) {
-                buffer2 += curString;
-            }
-        }
-        return buffer2;
-    }
-
-
-    private String addNumbersBuffer1(String curStr){
-
-        if (buffer1.equals("0") && (!dotSet1)) {
-            if (curStr.equals(".")) {
-                buffer1 += curStr;
-                dotSet1 = true;
-            } else {
-                buffer1 = curStr;
-            }
-        } else if (!buffer1.equals("0") && (!dotSet1)) {
-            if (curStr.equals(".")) {
-                buffer1 += curStr;
-                dotSet1 = true;
-            } else {
-                buffer1 += curStr;
-            }
-        } else if (!buffer1.equals("0") && dotSet1) {
-            if (!curStr.equals(".")) {
-                buffer1 += curStr;
-            }
-        }
-        return buffer1;
-    }
-
-    private String calculate(String buf1, String op, String buf2){
-
-        double result;
-        double num2;
-
-        double num1 = Double.parseDouble(buf1);
-        if (!buf2.equals("")) {
-            num2=Double.parseDouble(buf2);
-        } else {
-            //have no sense
-            //TODO: change
-            num2 = 0;
-        }
-
-        switch (op){
-            case "+":
-                result = num1 + num2;
-                break;
-            case "-":
-                result = num1 - num2;
-                break;
-            case "*":
-                result = num1*num2;
-                break;
-            case "/":
-                if (num2 == 0){
-                    return "error";
-                } else {
-                    result = num1 / num2;
-                    break;
-                }
-            default:
-                result = num1;
-        }
-
-        clearData();
-        return String.valueOf(result);
-    }
-
-    private void clearData(){
-        buffer1 = "0";
-        buffer2 = "";
-        operatorSet = false;
-        operator = "";
-        dotSet1 = false;
-        dotSet2 = false;
-        curResult = "";
-    }
-
-    private void copyCurResultToClipboard(String curResult){
-        Toolkit toolkit = Toolkit.getDefaultToolkit();
-        Clipboard clipboard = toolkit.getSystemClipboard();
-        StringSelection selection = new StringSelection(curResult);
-        clipboard.setContents(selection, selection);
-    }
 
     @Override
     public void keyTyped(KeyEvent e) {
         String s = String.valueOf(e.getKeyChar());
         if (operatorButtons.containsValue(s)){
-            actionsByButtons(s);
+            calculations.actionsByButtons(s);
         } else if (numberButtons.containsValue(s)){
-            actionsByButtons(s);
+            calculations.actionsByButtons(s);
         } if (s.equals("\n")){
-            actionsByButtons("=");
+            calculations.actionsByButtons("=");
         }
     }
 
@@ -360,5 +178,21 @@ public class Calculator extends JFrame implements ActionListener, KeyListener{
     @Override
     public void keyReleased(KeyEvent e) {
 
+    }
+
+    public JLabel getResultField() {
+        return resultField;
+    }
+
+    public HashMap<JButton, String> getNumberButtons() {
+        return numberButtons;
+    }
+
+    public HashMap<JButton, String> getOperatorButtons() {
+        return operatorButtons;
+    }
+
+    public JButton getClear() {
+        return clear;
     }
 }
